% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classify.R
\name{Classify}
\alias{Classify}
\title{Assign classificatin}
\usage{
Classify(classification, min_pvalue = 0.35)
}
\arguments{
\item{classification}{a list of ist samples pvalues for each subgroup}

\item{min_pvalue}{the lowest p-value that will be used to classify a sample (numeric value)}
}
\description{
Assign classificatin
}
