% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyEPN.R
\name{ClassifyEPN}
\alias{ClassifyEPN}
\title{Classify EPN bulk RNA-seq samples into eight (of the nine) molecular subgroups}
\usage{
ClassifyEPN(bulk, permutations = 10000)
}
\arguments{
\item{bulk}{a matrix of bulk RNA-seq sample (genes by samples) with common gene names (as row.names)}

\item{permutations}{the number of enrichment scores that will be calculated in order to estimate the sample distribution of enrichment scores}
}
\description{
Classify EPN bulk RNA-seq samples into eight (of the nine) molecular subgroups
}
