% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyPFA.R
\name{ClassifyPFA}
\alias{ClassifyPFA}
\title{Classify EPN bulk RNA-seq samples into PF A subgroups (PFA_1 and PFA_2)}
\usage{
ClassifyPFA(bulk, permutations = 10000)
}
\arguments{
\item{bulk}{a matrix of bulk RNA-seq sample (genes by samples) with common gene names (as row.names)}

\item{permutations}{the number of enrichment scores that will be calculated in order to estimate the sample distribution of enrichment scores}
}
\description{
Classify EPN bulk RNA-seq samples into PF A subgroups (PFA_1 and PFA_2)
}
