% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentScore.R
\name{EnrichmentScore}
\alias{EnrichmentScore}
\title{Calculate the enrichment score for a single bulk sample and gene set}
\usage{
EnrichmentScore(bulk, gene_set)
}
\arguments{
\item{bulk}{a matrix of a single bulk RNA-seq sample (genes by 1)}

\item{gene_set}{Character vector of genes}
}
\description{
Calculate the enrichment score for a single bulk sample and gene set
}
