% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentScore_EPN.R
\name{EnrichmentScore_EPN}
\alias{EnrichmentScore_EPN}
\title{Calculate the enrichment score
(with EPN gene set)}
\usage{
EnrichmentScore_EPN(bulk_sample)
}
\arguments{
\item{bulk_sample}{a matrix of a single bulk sample (genes by 1)}
}
\description{
Calculate the enrichment score
(with EPN gene set)
}
