% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentScore_PFA.R
\name{EnrichmentScore_PFA}
\alias{EnrichmentScore_PFA}
\title{Calculate the enrichment score
(with PFA gene set)}
\usage{
EnrichmentScore_PFA(bulk_sample)
}
\arguments{
\item{bulk_sample}{a matrix of a single bulk sample (genes by 1)}
}
\description{
Calculate the enrichment score
(with PFA gene set)
}
