% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindPvalue.R
\name{FindPvalue}
\alias{FindPvalue}
\title{Calculate the pvlaue associated with the enrichment score}
\usage{
FindPvalue(best_es, null_dist, gene_set)
}
\arguments{
\item{best_es}{list for each subgroup with its enrichment score located in "es"}

\item{null_dist}{list with an estimate of the sampling distribution of enrichment scores (null distribution)}

\item{gene_set}{character vector of genes}
}
\description{
Calculate the pvlaue associated with the enrichment score
}
