% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_EnrichmentScore.R
\name{internal_EnrichmentScore}
\alias{internal_EnrichmentScore}
\title{Calculate the enrichment score for a single bulk sample and gene set}
\usage{
internal_EnrichmentScore(num_bulk_genes, num_genes, indx_gene_set)
}
\arguments{
\item{num_bulk_genes}{number of genes the in bulk RNA-seq sample}

\item{num_genes}{number of genes in gene set}

\item{indx_gene_set}{numeric vector indexing the gene set in the bulk sample}
}
\description{
Calculate the enrichment score for a single bulk sample and gene set
}
