% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_distribution.R
\name{Make_null}
\alias{Make_null}
\title{Create null distribution of enrichment scores from random gene sets (estimate the enrichment score sampling distribution)}
\usage{
Make_null(permutations, bulk, num_genes)
}
\arguments{
\item{permutations}{the number of enrichment scores that will be calculated in order to estimate the sample distribution of enrichment scores}

\item{bulk}{a matrix of a single bulk RNA-seq sample (genes by 1)}

\item{num_genes}{the number of genes in the gene set}
}
\description{
Create null distribution of enrichment scores from random gene sets (estimate the enrichment score sampling distribution)
}
