% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_distribution.R
\name{null_EnrichmentScore}
\alias{null_EnrichmentScore}
\title{Enrichment score for gene_set in single bulk RNA-seq sample}
\usage{
null_EnrichmentScore(num_bulk_genes, num_genes, indx_gene_set)
}
\arguments{
\item{num_bulk_genes}{number of genes the in bulk RNA-seq sample}

\item{num_genes}{number of genes in gene set}

\item{indx_gene_set}{numeric vector indexing the gene set in genes the list of genes expressed in the bulk sample}
}
\description{
Enrichment score for gene_set in single bulk RNA-seq sample
}
