% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyEPN.R
\name{ClassifyEPN}
\alias{ClassifyEPN}
\title{Classify into eight (of the nine) molecular subgroups}
\usage{
ClassifyEPN(bulk, permutations = 10000)
}
\arguments{
\item{bulk}{a matrix of bulk transcriptomics data (genes by samples)}

\item{permutations}{the number permutations for a permutation test}
}
\description{
Classify into eight (of the nine) molecular subgroups
}
