% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyPFA.R
\name{ClassifyPFA}
\alias{ClassifyPFA}
\title{Classify into PF A subgroups (PFA_1 and PFA_2)}
\usage{
ClassifyPFA(bulk, permutations = 1e+05)
}
\arguments{
\item{bulk}{a matrix of bulk transcriptomics data (genes by samples)}

\item{permutations}{the number permutations for a permutation test}
}
\description{
Classify into PF A subgroups (PFA_1 and PFA_2)
}
